+++ HISTORY.rst
@@ -3,6 +3,11 @@
 History
 -------
 
+v0.18 (2019-08-30)
+..................
+* fix usage of ``max_burst_jobs``, improve coverage fix #152
+* stop lots of ``WatchVariableError`` errors in log, #153
+
 v0.17.1 (2019-08-21)
 ....................
 * deal better with failed job deserialization, #149 by @samuelcolvin